# Faster event processing when epoll transport is used