--- a/transport-native-epoll/src/main/java/io/netty/channel/epoll/AbstractEpollChannel.java
+++ b/transport-native-epoll/src/main/java/io/netty/channel/epoll/AbstractEpollChannel.java
@@ -38,7 +38,6 @@ abstract class AbstractEpollChannel extends AbstractChannel {
     protected int flags = Native.EPOLLET;
 
     protected volatile boolean active;
-    int id;
 
     AbstractEpollChannel(int fd, int flag) {
         this(null, fd, flag, false);