--- a/transport-native-epoll/src/main/java/io/netty/channel/epoll/EpollEventLoop.java
+++ b/transport-native-epoll/src/main/java/io/netty/channel/epoll/EpollEventLoop.java
@@ -50,11 +50,9 @@ final class EpollEventLoop extends SingleThreadEventLoop {
 
     private final int epollFd;
     private final int eventFd;
-    private final IntObjectMap<AbstractEpollChannel> ids = new IntObjectHashMap<AbstractEpollChannel>();
+    private final IntObjectMap<AbstractEpollChannel> channels = new IntObjectHashMap<AbstractEpollChannel>(4096);
     private final boolean allowGrowing;
-    private long[] events;
-
-    private int id;
+    private final EpollEventArray events;
 
     @SuppressWarnings("unused")
     private volatile int wakenUp;
@@ -64,10 +62,10 @@ final class EpollEventLoop extends SingleThreadEventLoop {
         super(parent, threadFactory, false);
         if (maxEvents == 0) {
             allowGrowing = true;
-            events = new long[128];
+            events = new EpollEventArray(4096);
         } else {
             allowGrowing = false;
-            events = new long[maxEvents];
+            events = new EpollEventArray(maxEvents);
         }
         boolean success = false;
         int epollFd = -1;
@@ -75,7 +73,7 @@ final class EpollEventLoop extends SingleThreadEventLoop {
         try {
             this.epollFd = epollFd = Native.epollCreate();
             this.eventFd = eventFd = Native.eventFd();
-            Native.epollCtlAdd(epollFd, eventFd, Native.EPOLLIN, 0);
+            Native.epollCtlAdd(epollFd, eventFd, Native.EPOLLIN);
             success = true;
         } finally {
             if (!success) {
@@ -97,29 +95,6 @@ final class EpollEventLoop extends SingleThreadEventLoop {
         }
     }
 
-    private int nextId() {
-        int id = this.id;
-        if (id == Integer.MAX_VALUE) {
-            // We used all possible ints in the past ( 1 - Integer.MAX_VALUE), time to scrub the stored channels
-            // and re-assign ids.
-            AbstractEpollChannel[] channels = ids.values(AbstractEpollChannel.class);
-            ids.clear();
-
-            id = 0;
-
-            for (AbstractEpollChannel ch: channels) {
-                id++;
-                ch.id = id;
-                ids.put(ch.id, ch);
-            }
-            if (id == Integer.MAX_VALUE) {
-                throw new IllegalStateException("Could not scrub ids");
-            }
-        }
-        this.id = ++id;
-        return id;
-    }
-
     @Override
     protected void wakeup(boolean inEventLoop) {
         if (!inEventLoop && WAKEN_UP_UPDATER.compareAndSet(this, 0, 1)) {
@@ -133,10 +108,9 @@ final class EpollEventLoop extends SingleThreadEventLoop {
      */
     void add(AbstractEpollChannel ch) {
         assert inEventLoop();
-        int id = nextId();
-        Native.epollCtlAdd(epollFd, ch.fd().intValue(), ch.flags, id);
-        ch.id = id;
-        ids.put(id, ch);
+        int fd = ch.fd().intValue();
+        Native.epollCtlAdd(epollFd, fd, ch.flags);
+        channels.put(fd, ch);
     }
 
     /**
@@ -144,7 +118,7 @@ final class EpollEventLoop extends SingleThreadEventLoop {
      */
     void modify(AbstractEpollChannel ch) {
         assert inEventLoop();
-        Native.epollCtlMod(epollFd, ch.fd().intValue(), ch.flags, ch.id);
+        Native.epollCtlMod(epollFd, ch.fd().intValue(), ch.flags);
     }
 
     /**
@@ -152,10 +126,14 @@ final class EpollEventLoop extends SingleThreadEventLoop {
      */
     void remove(AbstractEpollChannel ch) {
         assert inEventLoop();
-        if (ids.remove(ch.id) != null && ch.isOpen()) {
-            // Remove the epoll. This is only needed if it's still open as otherwise it will be automatically
-            // removed once the file-descriptor is closed.
-            Native.epollCtlDel(epollFd, ch.fd().intValue());
+
+        if (ch.isOpen()) {
+            int fd = ch.fd().intValue();
+            if (channels.remove(fd) != null) {
+                // Remove the epoll. This is only needed if it's still open as otherwise it will be automatically
+                // removed once the file-descriptor is closed.
+                Native.epollCtlDel(epollFd, ch.fd().intValue());
+            }
         }
     }
 
@@ -183,7 +161,7 @@ final class EpollEventLoop extends SingleThreadEventLoop {
         this.ioRatio = ioRatio;
     }
 
-    private int epollWait(boolean oldWakenUp) {
+    private int epollWait(boolean oldWakenUp) throws IOException {
         int selectCnt = 0;
         long currentTimeNanos = System.nanoTime();
         long selectDeadLineNanos = currentTimeNanos + delayNanos(currentTimeNanos);
@@ -275,9 +253,9 @@ final class EpollEventLoop extends SingleThreadEventLoop {
                     final long ioTime = System.nanoTime() - ioStartTime;
                     runAllTasks(ioTime * (100 - ioRatio) / ioRatio);
                 }
-                if (allowGrowing && ready == events.length) {
-                    // double the size of the array as we needed the whole space for the events
-                    events = new long[events.length << 1];
+                if (allowGrowing && ready == events.length()) {
+                    //increase the size of the array as we needed the whole space for the events
+                    events.increase();
                 }
                 if (isShuttingDown()) {
                     closeAll();
@@ -300,41 +278,52 @@ final class EpollEventLoop extends SingleThreadEventLoop {
     }
 
     private void closeAll() {
-        Native.epollWait(epollFd, events, 0);
-        Collection<AbstractEpollChannel> channels = new ArrayList<AbstractEpollChannel>(ids.size());
+        try {
+            Native.epollWait(epollFd, events, 0);
+        } catch (IOException ignore) {
+            // ignore on close
+        }
+        Collection<AbstractEpollChannel> array = new ArrayList<AbstractEpollChannel>(channels.size());
 
-        for (IntObjectMap.Entry<AbstractEpollChannel> entry: ids.entries()) {
-            channels.add(entry.value());
+        for (IntObjectMap.Entry<AbstractEpollChannel> entry: channels.entries()) {
+            array.add(entry.value());
         }
 
-        for (AbstractEpollChannel ch: channels) {
+        for (AbstractEpollChannel ch: array) {
             ch.unsafe().close(ch.unsafe().voidPromise());
         }
     }
 
-    private void processReady(long[] events, int ready) {
+    private void processReady(EpollEventArray events, int ready) {
         for (int i = 0; i < ready; i ++) {
-            final long ev = events[i];
-
-            int id = (int) (ev >> 32L);
-            if (id == 0) {
+            final int fd = events.fd(i);
+            if (fd == eventFd) {
                 // consume wakeup event
                 Native.eventFdRead(eventFd);
             } else {
-                AbstractEpollChannel ch = ids.get(id);
-                if (ch != null) {
+                final long ev = events.events(i);
+
+                AbstractEpollChannel ch = channels.get(fd);
+                if (ch != null && ch.isOpen()) {
+                    boolean close = (ev & Native.EPOLLRDHUP) != 0;
+                    boolean read = (ev & Native.EPOLLIN) != 0;
+                    boolean write = (ev & Native.EPOLLOUT) != 0;
+
                     AbstractEpollUnsafe unsafe = (AbstractEpollUnsafe) ch.unsafe();
-                    if ((ev & Native.EPOLLOUT) != 0 && ch.isOpen()) {
+                    if (write) {
                         // force flush of data as the epoll is writable again
                         unsafe.epollOutReady();
                     }
-                    if ((ev & Native.EPOLLIN) != 0 && ch.isOpen()) {
+                    if (read) {
                         // Something is ready to read, so consume it now
                         unsafe.epollInReady();
                     }
-                    if ((ev & Native.EPOLLRDHUP) != 0 && ch.isOpen()) {
+                    if (close) {
                         unsafe.epollRdHupReady();
                     }
+                } else {
+                    // We received an event for an fd which we not use anymore. Remove it from the epoll_event set.
+                    Native.epollCtlDel(epollFd, fd);
                 }
             }
         }
@@ -343,14 +332,19 @@ final class EpollEventLoop extends SingleThreadEventLoop {
     @Override
     protected void cleanup() {
         try {
-            Native.close(epollFd);
-        } catch (IOException e) {
-            logger.warn("Failed to close the epoll fd.", e);
-        }
-        try {
-            Native.close(eventFd);
-        } catch (IOException e) {
-            logger.warn("Failed to close the event fd.", e);
+            try {
+                Native.close(epollFd);
+            } catch (IOException e) {
+                logger.warn("Failed to close the epoll fd.", e);
+            }
+            try {
+                Native.close(eventFd);
+            } catch (IOException e) {
+                logger.warn("Failed to close the event fd.", e);
+            }
+        } finally {
+            // release native memory
+            events.free();
         }
     }
 }