@trustin 2015-02-07
Looks good to me. Please cherry-pick.