@normanmaurer 2015-02-08
cherry-picked into 4.0 , 4.1 and master